requirements.txt
asyncio
https://github.com/Rapptz/discord.py/archive/rewrite.zip#egg=discord.py
blessings
pyyaml
mistletoe == 0.6.2
